Math and Optional Parameters
Console app created with a class that has one required parameter and an optional parameter.  An user must enter the first number but not necessarily the second.  The class takes one or both parameters and returns the value of the math operation within.  The value is then displayed on the screen.