Update README.md
These are small projects I've developed to demonstrate my knowledge of the C# programming language and its components.  The C# language was used in conjunction with MS Visual Studio.  The .NET Framework using the MVC pattern design expanded my knowledge of Frameworks.  And then the Entity Frameworks with its code-first and database-first methods added more functionality to my projects.  Additionally, HTML and CSS was used on the .NET Web applications.